dagster-poc/pyproject.toml
[project]
name = "dagster-poc"
version = "0.1.0"
description = "Dagster + dbt Cloud orchestration monorepo"
requires-python = ">=3.10,<3.15"
dependencies = [
    "duckdb>=1.4.3",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "pytest>=7.0.0",
    "ruff",
]

[tool.uv]
package = false

[tool.uv.workspace]
members = [
    "ingestion",
    "dbt",
    "downstream",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = [
    "ingestion/tests",
    "dbt/tests",
    "downstream/tests",
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
